# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/event": {
      "post": {
        "operationId": "EventController_createEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "신규 이벤트 생성",
        "tags": [
          "이벤트 관련"
        ]
      },
      "get": {
        "operationId": "EventController_listEvents",
        "parameters": [
          {
            "name": "filterActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "filterType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "이벤트 목록 조회",
        "tags": [
          "이벤트 관련"
        ]
      }
    },
    "/event/{eventId}": {
      "get": {
        "operationId": "EventController_getEvent",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "이벤트 상세 조회",
        "tags": [
          "이벤트 관련"
        ]
      },
      "put": {
        "operationId": "EventController_updateEvent",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "이벤트 수정",
        "tags": [
          "이벤트 관련"
        ]
      },
      "delete": {
        "operationId": "EventController_deleteEvent",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "이벤트 삭제",
        "tags": [
          "이벤트 관련"
        ]
      }
    },
    "/reward": {
      "post": {
        "operationId": "RewardController_createReward",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRewardRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "신규 보상 생성",
        "tags": [
          "보상 관련"
        ]
      },
      "get": {
        "operationId": "RewardController_listReward",
        "parameters": [
          {
            "name": "filterActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "filterType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "보상 목록 조회",
        "tags": [
          "보상 관련"
        ]
      }
    },
    "/reward/{rewardId}": {
      "put": {
        "operationId": "RewardController_updateReward",
        "parameters": [
          {
            "name": "rewardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRewardRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "보상 수정",
        "tags": [
          "보상 관련"
        ]
      },
      "delete": {
        "operationId": "RewardController_deleteReward",
        "parameters": [
          {
            "name": "rewardID",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "allowed": [
              "operator"
            ]
          }
        ],
        "summary": "보상 삭제",
        "tags": [
          "보상 관련"
        ]
      }
    },
...
  },
  "info": {
    "title": "Event Manager",
    "description": "이벤트/보상 관리 시스템",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEventRequestDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "targetCriteria": {
            "type": "number"
          },
          "rewardIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "title",
          "targetCriteria",
          "rewardIds"
        ]
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "targetCriteria": {
            "type": "number"
          },
          "rewards": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "type",
          "title",
          "targetCriteria",
          "rewards",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
      "EventListDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "filterActive": {
            "type": "boolean"
          },
          "filterType": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "list"
        ]
      },
      "UpdateEventRequestDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "targetCriteria": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "rewardIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "active",
          "title",
          "targetCriteria",
          "rewardIds"
        ]
      },
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "result"
        ]
      },
      "CreateRewardRequestDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coupons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "point": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "title",
          "coupons",
          "items"
        ]
      },
      "RewardDto": {
        "type": "object",
        "properties": {
          "rewardId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coupons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "point": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "rewardId",
          "type",
          "title",
          "coupons",
          "items",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateRewardRequestDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coupons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "point": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "active",
          "title",
          "coupons",
          "items"
        ]
      },
      "RewardListDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "filterActive": {
            "type": "boolean"
          },
          "filterType": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RewardDto"
            }
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "list"
        ]
      },
...
    }
  }
}